bl-reload-gtk23: Load settings for xsettingsd from GTK3's settings.ini
After launching xsettingsd, xsettingsd takes full ownership of GTK3
settings such that any changes to settings.ini while xsettingsd is
running are not reflected in the properties of PyGObject's Gtk.Settings
class. This makes sense when in any desktop session, xsettingsd or
another daemon that implements the xsettingsd protocol like
gnome-settings-daemon is fully responsible for settings, and for keeping
settings.ini in sync with changes to the xsettings configuration (in
memory or on disk).

On BL, however, we use GTK3's settings.ini as the main interface for
changing theming information, so we need the changes to sync the other
way around (always from settings.ini to xsettingsd) and treat
settings.ini as authoritative.

This change makes GTK3's settings.ini the authoritative source for
theming information.